Moved actions and components into the $lib dir for easier importing
