+++ services/admin/src/lib/components/manifests/ManifestCanvasSelector.svelte
