+++ services/admin/src/lib/components/manifests/ManifestContentEditor.svelte
@@ -4,12 +4,12 @@
     CanvasManifest,
     Canvas,
   } from "@crkn-rcdr/access-data/src/access/CanvasManifest";
-  import Align from "../shared/Align.svelte";
-  import CanvasLabelEditor from "../canvases/CanvasLabelEditor.svelte";
-  import CanvasViewer from "../canvases/CanvasViewer.svelte";
-  import CanvasThumbnailList from "../canvases/CanvasThumbnailList.svelte";
-  import Switch from "../shared/Switch.svelte";
-  import SwitchCase from "../shared/SwitchCase.svelte";
+  import Align from "$lib/components/shared/Align.svelte";
+  import CanvasLabelEditor from "$lib/components/canvases/CanvasLabelEditor.svelte";
+  import CanvasViewer from "$lib/components/canvases/CanvasViewer.svelte";
+  import CanvasThumbnailList from "$lib/components/canvases/CanvasThumbnailList.svelte";
+  import Switch from "$lib/components/shared/Switch.svelte";
+  import SwitchCase from "$lib/components/shared/SwitchCase.svelte";
 
   export let manifest: CanvasManifest;
