+++ services/admin/src/lib/components/manifests/ManifestStatusIndicator.svelte
