+++ services/admin/src/lib/components/shared/SideMenuBody.svelte
