+++ services/admin/src/lib/components/shared/SideMenuPageListButton.svelte
