+++ services/admin/src/lib/components/shared/Switch.svelte
